With cast_xou function, fixed a part where it was converting the number wrongly with a l flag

+++ sources/caster.c
@@ -30,7 +30,7 @@ uintmax_t	cast_xou(t_data *info)
 	if (ft_strcmp(info->length_flags, "h") == 0)
 		number = (unsigned short)va_arg(info->args, unsigned int);
 	else if (ft_strcmp(info->length_flags, "hh") == 0)
-		number = (unsigned char)va_arg(info->args, long long);
+		number = (unsigned char)va_arg(info->args, unsigned int);
 	else if (ft_strcmp(info->length_flags, "l") == 0)
 		number = (unsigned long)va_arg(info->args, unsigned long int);
 	else if (ft_strcmp(info->length_flags, "ll") == 0)
